check file exist before remove

+++ vars/ubnt_builders.groovy
@@ -630,7 +630,9 @@ def debbox_builder(String productSeries, Map job_options=[:], Map build_series=[
                     // check SqaCI was existed
                     if(!fileExists("${sqaciDir}")) {
                         // always try to get latest version script
-                        sh "rm ${mkenv}"
+                        if(!fileExists("${mkenv}")) {
+                            sh "rm ${mkenv}"
+                        }                        
                         sh "wget http://tpe-pbsqa-ci.rad.ubnt.com:8888/share_space/SqaCI/mkenv.py -O ${mkenv}"
                         sh ". ~/.profile; python3 mkenv.py -ds -q -tp ${tokenPath}"
                         sh "${sqaCI} clone ${testCaseProj} -token ${tokenPath}"
